requirements.txt
amqp==5.0.6
asgiref==3.3.4
billiard==3.6.4.0
camelot-py==0.8.2
celery==5.0.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.7
detect-delimiter==0.1.1
Django==3.2.2
djangorestframework==3.12.4
et-xmlfile==1.1.0
kombu==5.1.0b1
numpy==1.20.3
opencv-python==4.5.2.52
openpyxl==3.0.7
pandas==1.2.4
pdfminer.six==20201018
prompt-toolkit==3.0.18
pycparser==2.20
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
xlrd==2.0.1
xlwt==0.7.2
